add page titles
