+++ src/pages/AddDoctorPage.jsx
@@ -1,5 +1,11 @@
+import { useDocumentTitle } from '@mantine/hooks';
+
 import AppShell from 'components/AppShell';
 
-const AddDoctorPage = () => <AppShell>AddDoctorPage</AppShell>;
+const AddDoctorPage = () => {
+  useDocumentTitle('Add Doctor - EasyClinic');
+
+  return <AppShell>AddDoctorPage</AppShell>;
+};
 
 export default AddDoctorPage;
