+++ src/pages/DeleteDoctorsPage.jsx
@@ -1,5 +1,11 @@
+import { useDocumentTitle } from '@mantine/hooks';
+
 import AppShell from 'components/AppShell';
 
-const DeleteDoctorsPage = () => <AppShell>DeleteDoctorsPage</AppShell>;
+const DeleteDoctorsPage = () => {
+  useDocumentTitle('Delete Doctors - EasyClinic');
+
+  return <AppShell>DeleteDoctorsPage</AppShell>;
+};
 
 export default DeleteDoctorsPage;
