+++ src/pages/DoctorDetailsPage.jsx
@@ -1,5 +1,11 @@
+import { useDocumentTitle } from '@mantine/hooks';
+
 import AppShell from 'components/AppShell';
 
-const DoctorDetailsPage = () => <AppShell>DoctorDetailsPage</AppShell>;
+const DoctorDetailsPage = () => {
+  useDocumentTitle('Doctor Details - EasyClinic');
+
+  return <AppShell>DoctorDetailsPage</AppShell>;
+};
 
 export default DoctorDetailsPage;
