+++ src/pages/MyAppointmentsPage.jsx
@@ -1,5 +1,11 @@
+import { useDocumentTitle } from '@mantine/hooks';
+
 import AppShell from 'components/AppShell';
 
-const MyAppointmentsPage = () => <AppShell>MyAppointmentsPage</AppShell>;
+const MyAppointmentsPage = () => {
+  useDocumentTitle('My Appointments - EasyClinic');
+
+  return <AppShell>MyAppointmentsPage</AppShell>;
+};
 
 export default MyAppointmentsPage;
