+++ src/pages/SigninPage.jsx
@@ -4,7 +4,7 @@ import AppShell from 'components/AppShell';
 import SigninForm from 'components/SigninForm';
 
 const SigninPage = () => {
-  useDocumentTitle('Sign in | EasyClinic');
+  useDocumentTitle('Sign in - EasyClinic');
 
   return (
     <AppShell>
