+++ src/pages/SignupPage.jsx
@@ -4,7 +4,7 @@ import AppShell from 'components/AppShell';
 import SignupForm from 'components/SignupForm';
 
 const SignupPage = () => {
-  useDocumentTitle('Sign up | EasyClinic');
+  useDocumentTitle('Sign up - EasyClinic');
 
   return (
     <AppShell>
